纠正返回的行数为整型
git-svn-id: http://dev.bjmaxinfo.com/svn/corner/trunk/corner@2097 bda6c913-8cf9-0310-89cc-b5856a464ced

+++ src/main/java/corner/orm/tapestry/validator/UniqueEntity.java
@@ -68,7 +68,7 @@ public class UniqueEntity extends BaseValidator {
 			return;
 		}
 		
-		long rowCount = (Long) ((HibernateDaoSupport) getEntityService()
+		int rowCount = (Integer) ((HibernateDaoSupport) getEntityService()
 				.getObjectRelativeUtils()).getHibernateTemplate().execute(
 				new HibernateCallback() {
 					public Object doInHibernate(Session session)
